+++ src/main/java/com/raoulvdberge/refinedstorage/container/ContainerFilter.java
@@ -1,8 +1,14 @@
 package com.raoulvdberge.refinedstorage.container;
 
 import com.raoulvdberge.refinedstorage.container.slot.filter.SlotFilter;
+import com.raoulvdberge.refinedstorage.container.slot.filter.SlotFilterFluid;
+import com.raoulvdberge.refinedstorage.inventory.fluid.FluidInventory;
+import com.raoulvdberge.refinedstorage.inventory.fluid.FluidInventoryFilter;
+import com.raoulvdberge.refinedstorage.inventory.fluid.FluidInventoryFilterIcon;
 import com.raoulvdberge.refinedstorage.inventory.item.ItemHandlerFilterIcon;
 import com.raoulvdberge.refinedstorage.inventory.item.ItemHandlerFilterItems;
+import com.raoulvdberge.refinedstorage.item.ItemFilter;
+import com.raoulvdberge.refinedstorage.tile.config.IType;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 
@@ -19,8 +25,12 @@ public class ContainerFilter extends ContainerBase {
 
         ItemHandlerFilterItems filter = new ItemHandlerFilterItems(stack);
 
+        FluidInventoryFilterIcon fluidIcon = new FluidInventoryFilterIcon(stack);
+        FluidInventory fluidFilter = new FluidInventoryFilter(stack);
+
         for (int i = 0; i < 27; ++i) {
-            addSlotToContainer(new SlotFilter(filter, i, x, y));
+            addSlotToContainer(new SlotFilter(filter, i, x, y).setEnableHandler(() -> ItemFilter.getType(stack) == IType.ITEMS));
+            addSlotToContainer(new SlotFilterFluid(fluidFilter, i, x, y).setEnableHandler(() -> ItemFilter.getType(stack) == IType.FLUIDS));
 
             if ((i + 1) % 9 == 0) {
                 x = 8;
@@ -30,11 +40,12 @@ public class ContainerFilter extends ContainerBase {
             }
         }
 
-        addSlotToContainer(new SlotFilter(new ItemHandlerFilterIcon(stack), 0, 8, 117));
+        addSlotToContainer(new SlotFilter(new ItemHandlerFilterIcon(stack), 0, 8, 117).setEnableHandler(() -> ItemFilter.getType(stack) == IType.ITEMS));
+        addSlotToContainer(new SlotFilterFluid(fluidIcon, 0, 8, 117).setEnableHandler(() -> ItemFilter.getType(stack) == IType.FLUIDS));
 
         addPlayerInventory(8, 149);
 
-        transferManager.addItemFilterTransfer(player.inventory, filter);
+        transferManager.addFilterTransfer(player.inventory, filter, fluidFilter, () -> ItemFilter.getType(stack));
     }
 
     public ItemStack getStack() {