+++ src/main/java/com/raoulvdberge/refinedstorage/gui/grid/filtering/GridFilterFilter.java
@@ -2,9 +2,12 @@ package com.raoulvdberge.refinedstorage.gui.grid.filtering;
 
 import com.raoulvdberge.refinedstorage.api.util.IFilter;
 import com.raoulvdberge.refinedstorage.apiimpl.API;
+import com.raoulvdberge.refinedstorage.gui.grid.stack.GridStackFluid;
 import com.raoulvdberge.refinedstorage.gui.grid.stack.GridStackItem;
 import com.raoulvdberge.refinedstorage.gui.grid.stack.IGridStack;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.fluids.FluidRegistry;
+import net.minecraftforge.fluids.FluidStack;
 
 import java.util.List;
 import java.util.function.Predicate;
@@ -18,10 +21,6 @@ public class GridFilterFilter implements Predicate<IGridStack> {
 
     @Override
     public boolean test(IGridStack stack) {
-        return accepts(filters, ((GridStackItem) stack).getStack(), stack.getModId());
-    }
-
-    public static boolean accepts(List<IFilter> filters, ItemStack stack, String stackModId) {
         if (filters.isEmpty()) {
             return true;
         }
@@ -31,12 +30,28 @@ public class GridFilterFilter implements Predicate<IGridStack> {
         for (IFilter filter : filters) {
             lastMode = filter.getMode();
 
-            if (filter.isModFilter()) {
-                if (filter.getStack().getItem().getRegistryName().getNamespace().equalsIgnoreCase(stackModId)) {
+            if (stack instanceof GridStackItem && filter.getStack() instanceof ItemStack) {
+                ItemStack stackInFilter = (ItemStack) filter.getStack();
+
+                if (filter.isModFilter()) {
+                    if (stackInFilter.getItem().getRegistryName().getNamespace().equalsIgnoreCase(stack.getModId())) {
+                        return filter.getMode() == IFilter.MODE_WHITELIST;
+                    }
+                } else if (API.instance().getComparer().isEqual(((GridStackItem) stack).getStack(), stackInFilter, filter.getCompare())) {
                     return filter.getMode() == IFilter.MODE_WHITELIST;
                 }
-            } else if (API.instance().getComparer().isEqual(stack, filter.getStack(), filter.getCompare())) {
-                return filter.getMode() == IFilter.MODE_WHITELIST;
+            } else if (stack instanceof GridStackFluid && filter.getStack() instanceof FluidStack) {
+                FluidStack stackInFilter = (FluidStack) filter.getStack();
+
+                if (filter.isModFilter()) {
+                    String stackInFilterModId = FluidRegistry.getModId(stackInFilter);
+
+                    if (stackInFilterModId != null && stackInFilterModId.equalsIgnoreCase(stack.getModId())) {
+                        return filter.getMode() == IFilter.MODE_WHITELIST;
+                    }
+                } else if (API.instance().getComparer().isEqual(((GridStackFluid) stack).getStack(), stackInFilter, filter.getCompare())) {
+                    return filter.getMode() == IFilter.MODE_WHITELIST;
+                }
             }
         }